# optimize FeatureDefinitionAnalyzer