--- a/tests/D2L.CodeStyle.Analyzers.Test/Specs/FeatureDefinitionAnalyzer.cs
+++ b/tests/D2L.CodeStyle.Analyzers.Test/Specs/FeatureDefinitionAnalyzer.cs
@@ -8,23 +8,29 @@ namespace SpecTests {
 	using System;
 	using D2L.LP.LaunchDarkly;
 
-	public sealed class ObjectDefinition :/* InvalidLaunchDarklyFeatureDefinition(long) */ FeatureDefinition<long> /**/{ }
+	public sealed class LongDefinition :/* InvalidLaunchDarklyFeatureDefinition(long) */ FeatureDefinition<long> /**/{ }
 	public sealed class ObjectDefinition :/* InvalidLaunchDarklyFeatureDefinition(System.TimeSpan) */ FeatureDefinition<TimeSpan> /**/{ }
 
-	public sealed class BoolDefinition : FeatureDefinition<bool> { }
-	public sealed class FloatDefinition : FeatureDefinition<float> { }
-	public sealed class IntDefinition : FeatureDefinition<int> { }
-	public sealed class StringDefinition : FeatureDefinition<string> { }
+	namespace Keywords {
+		public sealed class BoolDefinition : FeatureDefinition<bool> { }
+		public sealed class FloatDefinition : FeatureDefinition<float> { }
+		public sealed class IntDefinition : FeatureDefinition<int> { }
+		public sealed class StringDefinition : FeatureDefinition<string> { }
+	}
 
-	public sealed class BoolDefinition : FeatureDefinition<Boolean> { }
-	public sealed class FloatDefinition : FeatureDefinition<Single> { }
-	public sealed class IntDefinition : FeatureDefinition<Int32> { }
-	public sealed class StringDefinition : FeatureDefinition<String> { }
+	namespace SystemType {
+		public sealed class BoolDefinition : FeatureDefinition<Boolean> { }
+		public sealed class FloatDefinition : FeatureDefinition<Single> { }
+		public sealed class IntDefinition : FeatureDefinition<Int32> { }
+		public sealed class StringDefinition : FeatureDefinition<String> { }
+	}
 
-	public sealed class BoolDefinition : FeatureDefinition<System.Boolean> { }
-	public sealed class FloatDefinition : FeatureDefinition<System.Single> { }
-	public sealed class IntDefinition : FeatureDefinition<System.Int32> { }
-	public sealed class StringDefinition : FeatureDefinition<System.String> { }
+	namespace QualifiedSystemType {
+		public sealed class BoolDefinition : FeatureDefinition<System.Boolean> { }
+		public sealed class FloatDefinition : FeatureDefinition<System.Single> { }
+		public sealed class IntDefinition : FeatureDefinition<System.Int32> { }
+		public sealed class StringDefinition : FeatureDefinition<System.String> { }
+	}
 }
 
 namespace MissingLaunchDarklyUsingStatement {