With ImmutabilityAnalyzer as reference (given run variability)

| Analyzer | Before | After |
| -- | -- | -- |
| ImmutabilityAnalyzer | 12.200s | 11.892s |
| FeatureDefinitionAnalyzer | 1.598s | 13ms |